backend/requirements.txt
Django==4.2.1
django-cors-headers==4.2.0
django-filter==23.3
django-rest-auth==0.9.5
djangorestframework==3.14.0
